+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.1.3
+
+### New
+
+- The following input formats are now supported by default in addition to the original `YYYY-MM-DD`:
+`"MM-DD-YY", "MMM DD YYYY", "MMM D YYYY", "MMM D YY", "MMM DD YY"`
+
+## Fixes
+
+- A link to Pendulum docs is more specific now.
+
 ## 1.1.2
 
 ### New